+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix composer post update script.
 * Add a config parameter to choose who receives the notifyAccountCreated mail.
 * New «website title» field.
+* Safer YAML escape for Ansible class.
 
 ## v0.0.2